refactor: home per-node shower fields in shower_properties
shower_age (energy-aware t,s variant -> renamed shower_age_at_depth to
avoid colliding with the legacy fixed-1e8-GeV shower_age(T)),
greisen_particles, hillas_e2, and node_fields (renamed
shower_state_at_nodes -- it computes the per-node shower state, not
generic "fields") form a self-contained call-graph community with no
edges into the Cherenkov orchestration. Home them with the other
longitudinal-profile physics. Verbatim moves; goldens bit-identical.
